Allow setting servo independent of attaching/detaching.
The pin connected to the servo can now be managed without having to
attach to the servo first. The pin can also be kept for the ServoOsc
instance even after detaching.
This allows users of the library to have the ServoOsc instance save and
manage the pin independently of being attached/detached to/from the
servo, instead of having to keep a separate variable to manage the pin.

Updated all documentation and examples to use the new setPin(),
attach(), detach(), getPin(), isAttached(), etc. methods.