Fix for using multiple oscillators at the same.
Usage of a static variable in the `update()` member to keep track of the
last time the position was updated, meant that the variable was a class
variable and not an instance variable.
This caused erratic behavior if multiple instances were being run
concurrently.
Fixed it moving the variable to the private section.